AnalogMultiplexRead.ino
Some testbed code that will read up to eight pins of analog data from a CD4051 multiplexer.  In this case we have eight LM35DZ temperature sensors attached.

To do: Generalize it to up to 16 sensors.  Probably just make it two channels of 8bits since we have the pins available.